Refactor: 기기별 사이즈 대응 (#144, #146)
