+++ modules/post/components/detail/duration/Duration.tsx
@@ -4,7 +4,7 @@ import ScreenLayout from '@shared/components/layout/ScreenLayout';
 import { ScreenTitle } from '@shared/components/font/Font';
 import { View } from 'react-native';
 import DividerWrapper from '../../../../shared/components/layout/DividerWrapper';
-import ContentHeader from '../ContentHeader';
+import ContentHeader from '../../../../shared/components/layout/ContentHeader';
 import Schedule from './schedule/Schedule';
 import TimePickerButton from './timePickerButton/TimePickerButton';
 import { Iduration } from './type';
