+++ modules/post/components/detail/participants/Participants.tsx
@@ -6,7 +6,7 @@ import MemberCount from './memberCount/MemberCount';
 import Members from './members/Members';
 import { Iparticipants } from './type';
 import DividerWrapper from '../../../../shared/components/layout/DividerWrapper';
-import ContentHeader from '../ContentHeader';
+import ContentHeader from '../../../../shared/components/layout/ContentHeader';
 
 const TITLE = '모임 멤버';
