--- modules/post/components/detail/style.ts
@@ -1,5 +1,0 @@
-import { makeStyles } from '@rneui/themed';
-
-export default makeStyles(() => ({
-  cContentHeader: { flexDirection: 'row', marginTop: 28, marginBottom: 12 },
-}));
