--- modules/post/components/detail/type.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface IdividerWrapper {
-  children: ReactNode;
-}
