+++ modules/shared/components/layout/style.ts
@@ -2,12 +2,19 @@ import { makeStyles } from '@rneui/themed';
 import { applyRelativeSizes } from '@shared/utils/utils';
 
 export default makeStyles((theme, width: number) => {
-  const [DIVIDER_WIDTH, PADDING_VERTICAL, PADDING_HORIZONTAL] =
-    applyRelativeSizes({
-      width,
-      paddingVertical: 10,
-      paddingHorizontal: 15,
-    });
+  const [
+    DIVIDER_WIDTH,
+    PADDING_VERTICAL,
+    PADDING_HORIZONTAL,
+    CONTAINER_CONTENT_HEADER_MARGIN_TOP,
+    CONTAINER_CONTENT_HEADER_MARGIN_BOTTOM,
+  ] = applyRelativeSizes({
+    width,
+    paddingVertical: 10,
+    paddingHorizontal: 15,
+    marginTop: 28,
+    marginBottom: 12,
+  });
 
   return {
     dividerStyle: {
@@ -17,6 +24,11 @@ export default makeStyles((theme, width: number) => {
     container: {
       paddingVertical: PADDING_VERTICAL,
       paddingHorizontal: PADDING_HORIZONTAL,
+    },
+    cContentHeader: {
+      flexDirection: 'row',
+      marginTop: CONTAINER_CONTENT_HEADER_MARGIN_TOP,
+      marginBottom: CONTAINER_CONTENT_HEADER_MARGIN_BOTTOM,
     },
   };
 });
